requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
huggingface-hub>=0.20.0
transformers>=4.35.0
torch>=2.0.0
pygithub>=2.1.1
python-dotenv>=1.0.0
pydantic>=2.0.0
accelerate>=0.25.0

